PR-feedback: Lager interface av PdfClient istedet for PdfService
