+++ src/main/kotlin/no/nav/familie/baks/mottak/config/DokgenConfig.kt
@@ -3,14 +3,15 @@ package no.nav.familie.baks.mottak.config
 import no.nav.familie.baks.dokgen.DokGen
 import no.nav.familie.baks.mottak.config.featureToggle.FeatureToggleConfig
 import no.nav.familie.baks.mottak.config.featureToggle.UnleashNextMedContextService
+import no.nav.familie.baks.mottak.integrasjoner.DokgenPdfClient
 import no.nav.familie.baks.mottak.integrasjoner.FamilieDokumentClient
+import no.nav.familie.baks.mottak.integrasjoner.FamilieDokumentPdfClient
 import no.nav.familie.baks.mottak.integrasjoner.PdfClient
-import no.nav.familie.baks.mottak.søknad.GammelPdfService
-import no.nav.familie.baks.mottak.søknad.NyPdfService
-import no.nav.familie.baks.mottak.søknad.PdfService
-import no.nav.familie.baks.mottak.søknad.SøknadSpråkvelgerService
+import org.springframework.beans.factory.annotation.Qualifier
+import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
+import org.springframework.web.client.RestOperations
 
 @Configuration
 class DokgenConfig {
@@ -18,23 +19,22 @@ class DokgenConfig {
     fun dokgen() = DokGen()
 
     @Bean
-    fun pdfService(
+    fun pdfClient(
+        @Qualifier("restTemplateUnsecured") operations: RestOperations,
+        @Value("\${FAMILIE_BAKS_DOKGEN_API_URL}") dokgenUri: String,
         familieDokumentClient: FamilieDokumentClient,
-        søknadSpråkvelgerService: SøknadSpråkvelgerService,
-        dokgen: DokGen,
-        pdfClient: PdfClient,
+        dokGen: DokGen,
         unleashService: UnleashNextMedContextService,
-    ): PdfService {
+    ): PdfClient {
         if (unleashService.isEnabled(FeatureToggleConfig.BRUK_NY_DOKGEN_LØSNING)) {
-            return NyPdfService(
-                familieDokumentClient,
-                søknadSpråkvelgerService,
-                dokgen,
+            return DokgenPdfClient(
+                operations,
+                dokgenUri,
             )
         }
-        return GammelPdfService(
-            pdfClient,
-            søknadSpråkvelgerService,
+        return FamilieDokumentPdfClient(
+            familieDokumentClient,
+            dokGen,
         )
     }
 }
